Project 2 Commit
(Addresses issues #8 , #10 , #13 )
I haven't actually committed my work in a bit, so now's a good time. I got a bit caught up in chugging away and definitely didn't want to commit while the program was throwing a major "Memory is not a constructor" issue that made it totally suicide on startup. Here's a quick changelog:
- Added the Memory, MemoryAccessor, and MemoryManager files. Needs more work, but functional for now.
- Added linewrap functionality to the CLI. All the code for this is contained in Console.putText().
- Added buttons for activating and using the single-step CPU mode. The methods for accomplishing the new mode are all contained in control.ts.
- Added functionality to the load shell command. Not fully complete (does not assign an ID or create a PCB), but the command now acutally *loads*, which is pretty good for a "load" command.
- Added a few global constants that will hopefully make it easy to change around how memory works in the future and makes me much less reliant on constantly saying "_Memory.memory.length".
- CPU.ts is WIP.