Shuffling and Program Breaking
(Addresses issue #14 directly, others indirectly)
Made a few small cleanups. The main one is that large program loading and memory clearing is now done in memoryManager.ts, which is where it should be done instead of the shell. There is now also a program break option. The way I've coded it, it does NOT mark the PCB as completed. In fact, before shutting off the CPU execution, it saves the PCB in its current state. Due to how I set up my CPU's startup sequence, this means any program can be instantly restarted from right where it left off before it was broken with Ctrl+C.
Fun Changelog:
- devices.ts now passes event.ctrlKey as a parameter when sending a keyboard interrupt.
- Before checking for normal characters, deviceDriverKeyboard.ts now checks for 3 conditions: If Control is being held, if C is being pressed, and if the CPU is currently executing a program. Iff all 3 conditions are met, it will call the kernel's krnProgramBreak method. Otherwise, it will move on unfazed.
- krnShutdown now makes sure to turn off the CPU before ending.
- kernel.ts now has the krnProgramBreak method. This saves the currently executing process's PCB, then sets the CPU's exectuing flag to false.
- MemoryManager now does something, just to add another layer of safety.
- shell.ts's "load" command now uses the MemoryManager, rather than trying to do it all itself.