Displaying the Queue
Addresses issue #18. No need for a changelog here, I basically just added a button, a table, and a method to populate that table. The button toggles between the full PCB table and just the Ready List. Oh and, uh... I increased the width of all buttons by 30 pixels since a few couldnt fit all the text in them.