Backup Memory Protection and Question Answers
Though it likely won't happen, the Memory Manager now checks if the address is negative or above the segment size. If it is, it will ask the scheduler to end the current process and move on to whatever's next. The scheduler is also more verbose in the trace log.
More importantly, though, I added the answers to the lab questions. I hope you enjoy, because my hate for LaTeX grows by the minute.

Project 3 is now done.